// src/store/types/modalsTypes.ts
export enum types {
  AUTH_MODAL = 'AUTH_MODAL',
  CONFIRM_LOGOUT_MODAL = 'CONFIRM_LOGOUT_MODAL',
  CONFIRM_PROJECT_DELETE_MODAL = 'CONFIRM_PROJECT_DELETE_MODAL',
  CONFIRM_PROJECT_DELETE_MODAL_PROPS = 'CONFIRM_PROJECT_DELETE_MODAL_PROPS',

  CREATE_PROJECT_MODAL = 'CREATE_PROJECT_MODAL',
  CREATE_PROJECT_MODAL_PROPS = 'CREATE_PROJECT_MODAL_PROPS'
}

export type TTypes =
  types.AUTH_MODAL |
  types.CONFIRM_LOGOUT_MODAL |
  types.CONFIRM_PROJECT_DELETE_MODAL |
  types.CONFIRM_PROJECT_DELETE_MODAL_PROPS |
  types.CREATE_PROJECT_MODAL |
  types.CREATE_PROJECT_MODAL_PROPS